backend/requirements.txt
fastapi==0.122.0
python-socketio==5.10.0
uvicorn[standard]==0.38.0
openai==2.8.1
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23

# Authentication
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
